# app/Domain/Car/Models/Car.php
<?php

namespace App\Domain\Car\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Car extends Model
{
    use HasFactory;

    protected $fillable = ['brand', 'model', 'number', 'owner'];
}