# public/api.yaml
openapi: 3.0.0
info:
  title: lab3
  version: 1.0.0
servers:
  - url: http://u8200620-lab3.local/api/v1
    description: локальный сервер
paths:
  /cars:
    get:
      summary: получить список всех машин
      operationId: getCars
      responses:
        '200':
          description: массив пользователей
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarsBody"
        default:
          description: ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: создать машину
      operationId: createCar
      requestBody:
        description: данные для создания машины
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Car"
      responses:
        "201":
          description: объект создан
        default:
          description: ошибка
          $ref: "#/components/schemas/Error"
  /cars/{carId}:
    parameters:
      - description: id машины
        in: path
        required: true
        name: carId
        schema:
          type: integer
          format: int64
    get:
      summary: получить машину по id
      operationId: getCarById
      responses:
        '200':
          description: машина с заданным id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarBody"
    delete:
      summary: удалить машину по id
      operationId: deleteCarById
      responses:
        '200':
          description: машина удалена
    put:
      summary: заменить машину по id
      operationId: replaceCarById
      requestBody:
        description: объект машины для замены
        required: true
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/Car"
      responses:
        '200':
          description: машина заменена
    patch:
      summary: обновить заданные поля машины по id
      operationId: patchCarById
      requestBody:
        description: поля машины для замены
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Car"
      responses:
        '200':
          description: машина обновлена
  /car-services:
    get:
      summary: получить все автосервисы
      operationId: getAllCarServices
      responses:
        '200':
          description: массив автосервисов
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: "#/components/schemas/CarServicesBody"
    post:
      summary: создать автосервис
      operationId: createCarService
      requestBody:
        description: данные для создания автосервиса
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/CarService"
      responses:
        '200':
          description: автосервис создан
  /car-services/{serviceId}:
    parameters:
      - description: id автосервиса
        in: path
        required: true
        name: serviceId
        schema:
          type: integer
          format: int64
    get:
      summary: получить все автосервисы
      operationId: getCarServiceById
      responses:
        '200':
          description: массив автосервисов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarServiceBody"
    delete:
      summary: удалить автосервис по id
      operationId: deleteCarServiceById
      responses:
        '200':
          description: автосервис удален
    put:
      summary: заменить автосервис
      operationId: putCarServiceById
      requestBody:
        description: автосервис для замены
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarService"
      responses:
        '200':
          description: автосервис заменен
    patch:
      summary: обновить заданные поля автосервиса
      operationId: patchCarServiceById
      requestBody:
        description: поля автосервиса для обновления
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarService"
      responses:
        '200':
          description: автосервис обновлен
  /repairs:
    get:
      summary: получить все записи о ремонте
      operationId: getAllRepairs
      responses:
        '200':
          description: массив записей о ремонте
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: "#/components/schemas/RepairsBody"
    post:
      summary: создать запись о ремонте
      operationId: createCarService
      requestBody:
        description: данные для создания записи
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/Repair"
      responses:
        '200':
          description: запись создана
  /repairs/{repairId}:
    parameters:
      - description: id записи о ремонте
        in: path
        required: true
        name: repairId
        schema:
          type: integer
          format: int64
    get:
      summary: получить все записи о ремонте
      operationId: getRepairById
      responses:
        '200':
          description: массив записей
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepairBody"
    delete:
      summary: удалить запись о ремонте по id
      operationId: deleteRepairById
      responses:
        '200':
          description: запись удалена
    put:
      summary: заменить запись о ремонте
      operationId: putRepairById
      requestBody:
        description: запись для замены
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Repair"
      responses:
        '200':
          description: запись заменена
    patch:
      summary: обновить заданные поля записи о ремонте
      operationId: patchRepairById
      requestBody:
        description: поля записи для обновления
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Repair"
      responses:
        '200':
          description: запись обновлена

components:
  schemas:
    CarBody:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          $ref: "#/components/schemas/Car"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Error"
    CarsBody:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Car"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Error"
    CarServiceBody:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          $ref: "#/components/schemas/CarService"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Error"
    CarServicesBody:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CarService"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Error"
    RepairBody:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          $ref: "#/components/schemas/Repair"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Error"
    RepairsBody:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Repair"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Error"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string
        meta:
          type: object
    Car:
      type: object
      required:
        - id
        - brand
        - model
        - number
        - owner
      properties:
        id:
          type: integer
          format: int64
        brand:
          type: string
        model:
          type: string
        number:
          type: string
        owner:
          type: string
    CarService:
      type: object
      required:
        - id
        - name
        - address
        - owner
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        owner:
          type: string
    Repair:
      type: object
      required:
        - id
        - issue
        - price
        - car_id
        - car_service_id
      properties:
        id: 
          type: integer
          format: int64
        issue:
          type: string
        price:
          type: integer
          format: int32
        car_id: 
          type: integer
          format: int64
        car_service_id: 
          type: integer
          format: int64